requirements_dev.txt
alembic==1.12.0
anyio==3.6.2
asyncpg==0.27.0
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
click==8.1.3
databases==0.7.0
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
orjson==3.9.7
outcome==1.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pydantic==1.10.4
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-tornasync==0.6.0.post2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
tornado==6.3.3
trio==0.22.2
typing_extensions==4.4.0
ujson==5.8.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
